requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev ./pyproject.toml
#
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
black==25.11.0
    # via pacific_crest_trail (pyproject.toml)
build==1.3.0
    # via pip-tools
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
click==8.3.0
    # via
    #   black
    #   pip-tools
    #   uvicorn
click-spinner==0.1.10
    # via fastapi-utilities
croniter==1.4.1
    # via fastapi-utilities
cryptography==46.0.3
    # via pacific_crest_trail (pyproject.toml)
distlib==0.4.0
    # via virtualenv
fastapi==0.121.1
    # via
    #   fastapi-utilities
    #   pacific_crest_trail (pyproject.toml)
fastapi-utilities==0.3.1
    # via pacific_crest_trail (pyproject.toml)
filelock==3.20.0
    # via virtualenv
flake8==7.3.0
    # via pacific_crest_trail (pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via pacific_crest_trail (pyproject.toml)
identify==2.6.15
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   httpx
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pip-tools==7.5.1
    # via pacific_crest_trail (pyproject.toml)
platformdirs==4.5.0
    # via
    #   black
    #   virtualenv
pre-commit==4.4.0
    # via pacific_crest_trail (pyproject.toml)
pycodestyle==2.14.0
    # via flake8
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   fastapi-utilities
    #   pacific_crest_trail (pyproject.toml)
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via pacific_crest_trail (pyproject.toml)
pyflakes==3.4.0
    # via flake8
pyjwt==2.10.1
    # via pacific_crest_trail (pyproject.toml)
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via croniter
python-dotenv==1.2.1
    # via pydantic-settings
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via pre-commit
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   fastapi-utilities
    #   sqlmodel
sqlmodel==0.0.27
    # via pacific_crest_trail (pyproject.toml)
starlette==0.49.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.38.0
    # via pacific_crest_trail (pyproject.toml)
virtualenv==20.35.4
    # via pre-commit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
